feat: add course section in home page
